% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.conn.metrics.R
\name{get.conn.metrics}
\alias{get.conn.metrics}
\title{Get connectivity metrics for steady states}
\usage{
get.conn.metrics(suit.con.PA = inter.ep.in, nnet = net, shape.PA = all.ep)
}
\arguments{
\item{suit.con.PA}{shapefile of suitable habitat within ecological continuities within protected areas (i.e. suitable habitat connected and protected).}

\item{nnet}{network of potential connections among protected areas.}

\item{shape.PA}{shapefile of protected areas including a SITECODE column.}
}
\value{
a list with local (i.e., for each network node) and global (i.e. for the whole network) PC metrics calculated together with node betweenness.
}
\description{
get.conn.metrics calculates connectivity metrics (probability of connectivity - PC, and betweenness) for a network.
}
