% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.dist.R
\name{get.dist}
\alias{get.dist}
\title{Distance extraction}
\usage{
get.dist(df, mat.dist = ep.dist)
}
\arguments{
\item{df}{a one row data.frame containing a column from and a column to (network edge).}

\item{mat.dist}{a matrix of distance. Row and column names of mat.dist should match sitenames in df.}
}
\value{
A one row data.frame of distance.
}
\description{
get.dist extracts the distance between two sites from a distance matrix.
}
