% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.edgelst.R
\name{get.edgelst}
\alias{get.edgelst}
\title{Get an edgelist}
\usage{
get.edgelst(df)
}
\arguments{
\item{df}{a data.frame containing at least a column named SITECODE indicating site ID.}
}
\value{
A data.frame with all possible pairwise site connections (columns from and to) that can be used as an edgelist to generate a network of connections.
}
\description{
get.edgelst generates all possible pairwise site connections from a vector of sites.
}
