% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFLC-SCP_FutureProjections-package.R
\docType{package}
\name{RFLC-SCP_FutureProjections-package}
\alias{RFLC-SCP_FutureProjections}
\alias{RFLC-SCP_FutureProjections-package}
\title{RFLC-SCP_FutureProjections: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mcpri3/RFLC-SCP_FutureProjections}
  \item Report bugs at \url{https://github.com/mcpri3/RFLC-SCP_FutureProjections/issues}
}

}
\author{
\strong{Maintainer}: Marie-Caroline Prima \email{marie-caroline.prima@univ-grenoble-alpes.fr} (\href{https://orcid.org/{{orcid}}}{ORCID}) [copyright holder]

}
\keyword{internal}
